terminando y arreglando la carga desde excel

+++ ClienteV3/src/java/Converters/MaterialesController.java
@@ -135,7 +135,7 @@ public class MaterialesController implements Serializable {
             return key;
         }
 
-        String getStringKey(java.lang.Integer value) {
+        String getStringKey(java.lang.Long value) {
             StringBuilder sb = new StringBuilder();
             sb.append(value);
             return sb.toString();
